upds

+++ OCCPort/TColStd_Array1OfReal.cs
@@ -1,6 +1,6 @@
 ﻿namespace OCCPort
 {
-    internal class TColStd_Array1OfReal : NCollection_Array1<float>
+    internal class TColStd_Array1OfReal : NCollection_Array1<double>
     {
         public TColStd_Array1OfReal(int v1, int v2) : base(v1, v2)
         {
